refactor objectives context
