+++ lib/cambiatus_web/resolvers/commune.ex
@@ -3,11 +3,9 @@ defmodule CambiatusWeb.Resolvers.Commune do
   This module holds the implementation of the resolver for the commune context
   use this to resolve any queries and mutations for the Commune context
   """
-  alias Absinthe.Relay.Connection
 
   alias Cambiatus.{Auth, Commune, Error, Shop, Social}
   alias Cambiatus.Commune.Community
-  alias Cambiatus.Objectives.Claim
 
   def search(_, %{community_id: symbol}, _) do
     Commune.get_community(symbol)
@@ -19,94 +17,6 @@ defmodule CambiatusWeb.Resolvers.Commune do
   @spec get_transfer(map(), map(), map()) :: {:ok, Transfer.t()} | {:error, term}
   def get_transfer(_, %{id: id}, _) do
     Commune.get_transfer(id)
-  end
-
-  @doc """
-  Fetches a claim
-  """
-  @spec get_claim(map(), map(), map()) :: {:ok, Claim.t()} | {:error, term}
-  def get_claim(_, %{id: id}, _) do
-    Commune.get_claim(id)
-  end
-
-  def get_analyzed_claims(
-        _,
-        %{community_id: community_id} = args,
-        %{context: %{current_user: current_user}}
-      ) do
-    query = Commune.analyzed_claims_query(community_id, current_user.account)
-    count = Claim.count(query)
-
-    query =
-      if Map.has_key?(args, :filter) do
-        args
-        |> Map.get(:filter)
-        |> Enum.reduce(query, fn
-          {:claimer, claimer}, query ->
-            Claim.with_claimer(query, claimer)
-
-          {:status, status}, query ->
-            Claim.with_status(query, status)
-
-          {:direction, direction}, query ->
-            Claim.ordered(query, direction)
-        end)
-      else
-        Claim.ordered(query, :desc)
-      end
-
-    query
-    |> Connection.from_query(&Cambiatus.Repo.all/1, args)
-    |> case do
-      {:ok, result} ->
-        {:ok, Map.put(result, :count, Cambiatus.Repo.one(count))}
-
-      default ->
-        default
-    end
-  end
-
-  def get_pending_claims(_, %{community_id: community_id} = args, %{
-        context: %{current_user: current_user}
-      }) do
-    query = Commune.pending_claims_query(community_id, current_user.account)
-    count = Claim.count(query)
-
-    query =
-      if Map.has_key?(args, :filter) do
-        args
-        |> Map.get(:filter)
-        |> Enum.reduce(query, fn
-          {:claimer, claimer}, query ->
-            Claim.with_claimer(query, claimer)
-
-          {:status, status}, query ->
-            Claim.with_status(query, status)
-
-          {:direction, direction}, query ->
-            Claim.ordered(query, direction)
-        end)
-      else
-        Claim.ordered(query, :desc)
-      end
-
-    query
-    |> Connection.from_query(&Cambiatus.Repo.all/1, args)
-    |> case do
-      {:ok, result} ->
-        {:ok, Map.put(result, :count, Cambiatus.Repo.one(count))}
-
-      default ->
-        default
-    end
-  end
-
-  @doc """
-  Fetch a objective
-  """
-  @spec get_objective(map(), map(), map()) :: {:ok, map()} | {:error, term}
-  def get_objective(_, params, _) do
-    Commune.get_objective(params.id)
   end
 
   @doc """
@@ -169,15 +79,6 @@ defmodule CambiatusWeb.Resolvers.Commune do
     Commune.get_transfer_count(community)
   end
 
-  @spec get_action_count(Cambiatus.Commune.Community.t(), any, any) :: {:ok, any}
-  def get_action_count(%Community{} = community, _, _) do
-    Commune.get_action_count(community)
-  end
-
-  def get_claim_count(%Community{} = community, _, _) do
-    Commune.get_claim_count(community)
-  end
-
   def get_product_count(%Community{} = community, _, _) do
     Shop.community_product_count(community.symbol)
   end
@@ -190,13 +91,6 @@ defmodule CambiatusWeb.Resolvers.Commune do
   @spec get_invitation(map(), map(), map()) :: {:ok, list(map())} | {:error, String.t()}
   def get_invitation(_, %{id: id}, _) do
     Auth.get_invitation(id)
-  end
-
-  @spec complete_objective(map(), map(), map()) :: {:ok, map()} | {:error, String.t()}
-  def complete_objective(_, %{input: %{objective_id: id}}, %{
-        context: %{current_user: current_user}
-      }) do
-    Commune.complete_objective(current_user, id)
   end
 
   def domain_available(_, %{domain: domain}, _) do
