+++ lib/cambiatus_web/schema.ex
@@ -7,7 +7,7 @@ defmodule CambiatusWeb.Schema do
   use Absinthe.Schema
   use Absinthe.Relay.Schema, :classic
 
-  alias Cambiatus.{Accounts, Commune, Kyc, Payments, Shop, Social}
+  alias Cambiatus.{Accounts, Commune, Kyc, Payments, Shop, Social, Objectives}
 
   def context(ctx) do
     loader =
@@ -18,6 +18,7 @@ defmodule CambiatusWeb.Schema do
       |> Dataloader.add_source(Shop, Shop.data())
       |> Dataloader.add_source(Social, Social.data())
       |> Dataloader.add_source(Payments, Payments.data())
+      |> Dataloader.add_source(Objectives, Objectives.data())
 
     Map.put(ctx, :loader, loader)
   end
